Added GUI base on wxWidget
Added a project4 directory that contains a VS sln to compile a
testRequest GUI base on wxWidget. Path to wx/wx.h must be
included correctly after extracting the SDK and compiling the libs.
I have added a link to a tutorial on getting started on wxWidget
within the README.